refactor(client): client 중복 삭제
